Errors corrected
The name for class DataServices was changed to DataService, and I've corrected several typo.
The special method for converting str into datetime format was written in trip_data_reader.py.
Since there was no need in creating configuration module in the task, and amount of data files is not that big, I didn't keep the configuration.
Nevertheless, now I am aware of what is it and will use configuration files in future projects.